+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/omBarnetFormComponents.tsx
@@ -1,0 +1,4 @@
+import { getTypedFormComponents, ValidationError } from '@navikt/sif-common-formik-ds';
+import { OmBarnetFormFields, OmBarnetFormValues } from './OmBarnetStep';
+
+export const omBarnetFormComponents = getTypedFormComponents<OmBarnetFormFields, OmBarnetFormValues, ValidationError>();
