+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/spørsmål/HøyereRisikoForFraværSpørsmål.tsx
@@ -1,0 +1,20 @@
+import { getYesOrNoValidator } from '@navikt/sif-common-formik-ds/src/validation';
+import { useAppIntl } from '../../../../i18n';
+import { omBarnetFormComponents } from '../omBarnetFormComponents';
+import { OmBarnetFormFields } from '../OmBarnetStep';
+
+const { YesOrNoQuestion } = omBarnetFormComponents;
+
+const HøyereRisikoForFraværSpørsmål = () => {
+    const { text } = useAppIntl();
+    return (
+        <YesOrNoQuestion
+            name={OmBarnetFormFields.høyereRisikoForFravær}
+            legend={text('steg.omBarnet.spm.høyereRisikoForFravær')}
+            data-testid="høyereRisikoForFravær"
+            validate={getYesOrNoValidator()}
+        />
+    );
+};
+
+export default HøyereRisikoForFraværSpørsmål;
