+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/spørsmål/KroniskEllerFunksjonshemningSpørsmål.tsx
@@ -1,0 +1,19 @@
+import { getYesOrNoValidator } from '@navikt/sif-common-formik-ds/src/validation';
+import { useAppIntl } from '../../../../i18n';
+import { omBarnetFormComponents } from '../omBarnetFormComponents';
+import { OmBarnetFormFields } from '../OmBarnetStep';
+
+const { YesOrNoQuestion } = omBarnetFormComponents;
+
+const KroniskEllerFunksjonshemningSpørsmål = () => {
+    const { text } = useAppIntl();
+    return (
+        <YesOrNoQuestion
+            name={OmBarnetFormFields.kroniskEllerFunksjonshemming}
+            legend={text('steg.omBarnet.spm.kroniskEllerFunksjonshemmende')}
+            validate={getYesOrNoValidator()}
+        />
+    );
+};
+
+export default KroniskEllerFunksjonshemningSpørsmål;
